project_demo_backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-paypal==2.1
django-taggit==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
